deploy.bash safety measures added a confirmation step into deploy.bash itself, and updated the readme to warn about using it too often

+++ deploy.bash
@@ -1,15 +1,21 @@
 #!/bin/bash
 
-# build front-end
-cd front
-npm run build
+# confirm that user really means to do this
+echo "Do you really mean to deploy? This shouldn't be done often. Enter \"yes\" to confirm."
+read CONFIRMATION
+if [ $CONFIRMATION == "yes" ]
+then
+    # build front-end
+    cd front
+    npm run build
 
-# put built front-end into the django app
-cd ..
-rm -rf ./server/front-build
-mv ./front/build ./server/front-build
+    # put built front-end into the django app
+    cd ..
+    rm -rf ./server/front-build
+    mv ./front/build ./server/front-build
 
-# push changes to the repo
-git add .
-git commit -m "updated build"
-git push origin master
+    # push changes to the repo
+    git add .
+    git commit -m "updated build"
+    git push origin master
+fi
